fastapi/requirements.txt
alembic==1.9.4
anyio==3.6.2
fastapi==0.92.0
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
pydantic==1.10.5
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
typing_extensions==4.5.0